+++ starsky-tools/end2end/cypress/plugins/index.ts
@@ -14,6 +14,7 @@ module.exports = (_, config) => {
     console.log('ignored due existing env names =>', process.env.CYPRESS_BASE_URL, process.env.cypress_name, process.env.cypress_AUTH_USER, 'cypress_AUTH_PASS')
     return
   }
+  console.log('running file based settings')
 
   // accept a configEnv value or use development by default
   const file = config.env.configEnv || 'local'