+++ starsky/starsky.foundation.storage/Storage/StorageHostFullPathFilesystem.cs
@@ -4,6 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using starsky.foundation.injection;
+using starsky.foundation.platform.Helpers;
+using starsky.foundation.platform.Interfaces;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Models;
 
@@ -12,6 +14,12 @@ namespace starsky.foundation.storage.Storage
 	[Service(typeof(IStorage), InjectionLifetime = InjectionLifetime.Scoped)]
 	public class StorageHostFullPathFilesystem : IStorage
 	{
+		private readonly IWebLogger _logger;
+
+		public StorageHostFullPathFilesystem(IWebLogger logger = null)
+		{
+			_logger = logger;
+		}
 
 		/// <summary>
 		/// Get the storage info
@@ -52,12 +60,14 @@ namespace starsky.foundation.storage.Storage
 			{
 				Directory.Delete(path, true);
 			}
-			catch (IOException) 
+			catch (IOException exception) 
 			{
+				_logger?.LogInformation(exception, "[FolderDelete] catch-ed IOException");
 				Directory.Delete(path, true);
 			}
-			catch (UnauthorizedAccessException)
+			catch (UnauthorizedAccessException exception)
 			{
+				_logger?.LogInformation(exception, "[FolderDelete] catch-ed UnauthorizedAccessException");
 				Directory.Delete(path, true);
 			}
 			return true;
@@ -77,7 +87,7 @@ namespace starsky.foundation.storage.Storage
 			}
 			catch ( UnauthorizedAccessException e )
 			{
-				Console.WriteLine($"catch-ed UnauthorizedAccessException{e.Message}");
+				_logger?.LogError(e, "[GetAllFilesInDirectory] catch-ed UnauthorizedAccessException");
 				return new string[]{};
 			}
 
@@ -135,7 +145,7 @@ namespace starsky.foundation.storage.Storage
 				}
 				catch(UnauthorizedAccessException e) 
 				{
-					Console.WriteLine("Catch-ed UnauthorizedAccessException => " + e.Message);
+					_logger?.LogError("Catch-ed UnauthorizedAccessException => " + e.Message);
 				}
 			}
 			return folderList.OrderBy(p => p);
@@ -161,7 +171,7 @@ namespace starsky.foundation.storage.Storage
 			}
 			catch ( FileNotFoundException e)
 			{
-				Console.WriteLine(e);
+				_logger?.LogError(e, "[ReadStream] catch-ed FileNotFoundException");
 				return Stream.Null;
 			}
 			return fileStream;
@@ -185,7 +195,7 @@ namespace starsky.foundation.storage.Storage
 		}
 
 		/// <summary>
-		/// is the subpath a folder or file, or deleted (FolderOrFileModel.FolderOrFileTypeList.Deleted)
+		/// is the subPath a folder or file, or deleted (FolderOrFileModel.FolderOrFileTypeList.Deleted)
 		/// </summary>
 		/// <param name="path">fullFilePath</param>
 		/// <returns>is file, folder or deleted</returns>
@@ -239,27 +249,37 @@ namespace starsky.foundation.storage.Storage
 		public bool FileDelete(string path)
 		{
 			if ( !File.Exists(path) ) return false;
-			File.Delete(path);
-			return true;
+			bool LocalRun()
+			{
+				File.Delete(path);
+				return true;
+			}
+			return RetryHelper.Do(LocalRun, TimeSpan.FromSeconds(1));
 		}
 
 		public bool WriteStream(Stream stream, string path)
 		{
 			if ( !stream.CanRead ) return false;
 
-			stream.Seek(0, SeekOrigin.Begin);
-			
-			using (var fileStream = new FileStream(path, 
-				FileMode.Create, 
-				FileAccess.Write,FileShare.ReadWrite,
-				4096, 
-				FileOptions.Asynchronous))
+			bool LocalRun()
 			{
-				stream.CopyTo(fileStream);
-			}
+				stream.Seek(0, SeekOrigin.Begin);
+			
+				using (var fileStream = new FileStream(path, 
+					FileMode.Create, 
+					FileAccess.Write,FileShare.ReadWrite,
+					4096, 
+					FileOptions.Asynchronous))
+				{
+					stream.CopyTo(fileStream);
+					fileStream.Dispose();
+				}
 
-			stream.Dispose();
-			return true;
+				stream.Dispose();
+				return true;
+			}
+						
+			return RetryHelper.Do(LocalRun, TimeSpan.FromSeconds(1));
 		}
 
 		public bool WriteStreamOpenOrCreate(Stream stream, string path)
@@ -288,14 +308,21 @@ namespace starsky.foundation.storage.Storage
 		public async Task<bool> WriteStreamAsync(Stream stream, string path)
 		{
 			if ( !stream.CanRead ) return false;
-			using (var fileStream = new FileStream(path, FileMode.Create, 
-				FileAccess.Write, FileShare.Read, 4096, 
-				FileOptions.Asynchronous | FileOptions.SequentialScan))
+
+			async Task<bool> LocalRun()
 			{
-				await stream.CopyToAsync(fileStream);
+				using (var fileStream = new FileStream(path, FileMode.Create, 
+					FileAccess.Write, FileShare.Read, 4096, 
+					FileOptions.Asynchronous | FileOptions.SequentialScan))
+				{
+					await stream.CopyToAsync(fileStream);
+					fileStream.Dispose();
+				}
+				stream.Dispose();
+				return true;
 			}
-			stream.Dispose();
-			return true;
+
+			return await RetryHelper.DoAsync(LocalRun, TimeSpan.FromSeconds(1));
 		}
 
 		/// <summary>
@@ -305,17 +332,17 @@ namespace starsky.foundation.storage.Storage
 		/// <returns></returns>
 		public IEnumerable<string> GetAllFilesInDirectoryRecursive(string path)
         {
-            List<string> findlist = new List<string>();
+            List<string> findList = new List<string>();
 
             /* I begin a recursion, following the order:
              * - Insert all the files in the current directory with the recursion
              * - Insert all subdirectories in the list and re-begin the recursion from there until the end
              */
-            RecurseFind( path, findlist );
+            RecurseFind( path, findList );
 
             // Add filter for file types
             var imageFilesList = new List<string>();
-            foreach (var file in findlist)
+            foreach (var file in findList)
             {
 	            imageFilesList.Add(file);
             }